Use file descriptor of device for generic ethernet type
When constructing the QEMU network arguments of generic ethernet
type, if the script option is not specified or is empty, pass in
the file descriptor of the interface instead of its name.
It is preferable to manage the life cycle of the network
inteface in libvirt because not doing so forces libvirt to
delegate this responsibility to QEMU, and as an undesirable
consequence, libvirt must grant QEMU privileged access on the
host to configure the interface.  Since QEMU accepts file
descriptors so long as no device script is also provided in the
command, similarly when there is no script specified, libvirt
should pass in the file descriptors to QEMU.